Extract CompositionRoot to class
