+++ source/B2B.Transactions/B2B.Transactions.csproj
@@ -49,4 +49,8 @@ limitations under the License.
       </Reference>
     </ItemGroup>
 
+    <ItemGroup>
+      <Compile Remove="Infrastructure\CompositionRoot.cs" />
+    </ItemGroup>
+
 </Project>
