+++ source/Energinet.DataHub.MarketRoles.sln
@@ -56,6 +56,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "B2B.Transactions.Tests", "B
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "B2B.Transactions", "B2B.Transactions\B2B.Transactions.csproj", "{FB4D9AE5-FCF2-442D-97EC-8AED1F9DDD2E}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "B2B.Transactions.Infrastructure", "B2B.Transactions.Infrastructure\B2B.Transactions.Infrastructure.csproj", "{8B1DCD6E-013F-4B45-B9EB-167F1D874793}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -134,6 +136,10 @@ Global
 		{FB4D9AE5-FCF2-442D-97EC-8AED1F9DDD2E}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{FB4D9AE5-FCF2-442D-97EC-8AED1F9DDD2E}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{FB4D9AE5-FCF2-442D-97EC-8AED1F9DDD2E}.Release|Any CPU.Build.0 = Release|Any CPU
+		{8B1DCD6E-013F-4B45-B9EB-167F1D874793}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8B1DCD6E-013F-4B45-B9EB-167F1D874793}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8B1DCD6E-013F-4B45-B9EB-167F1D874793}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8B1DCD6E-013F-4B45-B9EB-167F1D874793}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{654DA496-60D8-47D8-923E-F24F71A31F84} = {FD29AEE1-CB22-41FD-A875-1E99C4D5D81C}
@@ -149,5 +155,6 @@ Global
 		{B14EBA74-CCAF-4677-8AE5-A0F6302ED54C} = {1C877C68-8E3D-4371-B96A-7360850A538E}
 		{76FA7684-27E1-4022-9FB3-F72811E45C60} = {1C877C68-8E3D-4371-B96A-7360850A538E}
 		{FB4D9AE5-FCF2-442D-97EC-8AED1F9DDD2E} = {1C877C68-8E3D-4371-B96A-7360850A538E}
+		{8B1DCD6E-013F-4B45-B9EB-167F1D874793} = {1C877C68-8E3D-4371-B96A-7360850A538E}
 	EndGlobalSection
 EndGlobal
